Fixed for position calculation

+++ BLE/CalUnit.cs
@@ -9,8 +9,8 @@ namespace BLE
     class CalUnit
     {
         public double[] Position { get; set; } = new double[3];
-        public float Signal { get; set; }
-        public CalUnit(double[] position, float signal)
+        public double Signal { get; set; }
+        public CalUnit(double[] position, double signal)
         {
             this.Signal = signal;
             this.Position = position;
